Adapt the CalendarView accordingly and use calendar API endpoint
