+++ flirror/api.py
@@ -28,3 +28,24 @@ class WeatherApi(FlirrorMethodView):
         # TODO This should be done before storing the data, but I'm not sure
         # how to tell Pony how to serialize the datetime to JSON
         return weather
+
+
+class CalendarApi(FlirrorMethodView):
+
+    endpoint = "api-calendar"
+    rule = "/api/calendar"
+    template_name = None
+
+    FLIRROR_OBJECT_KEY = "module_calendar"
+
+    def get(self):
+        data = self.get_events()
+        return jsonify(data)
+
+    def get_events(self):
+        db = current_app.extensions["database"]
+        data = get_object_by_key(db, self.FLIRROR_OBJECT_KEY)
+        # Change timestamps to datetime objects
+        # TODO This should be done before storing the data, but I'm not sure
+        # how to tell Pony how to serialize the datetime to JSON
+        return data
