+++ flirror/templates/calendar.html
@@ -7,18 +7,18 @@
         {% for item in events %}
         <div class="row mb-2">
             <div class="col-3 text-right">
-                <h2 class="font-weight-light"><strong>{{ item.start.strftime("%d") }}</strong></h2>
-                <h6 class="text-uppercase">{{ item.start.strftime("%b") }}</h6>
+                <h2 class="font-weight-light"><strong>{{ item.start | format_time("%d") }}</strong></h2>
+                <h6 class="text-uppercase">{{ item.start | format_time("%b") }}</h6>
             </div>
             <div class="col-9 align-self-center">
                 <h6 class="text-uppercase"><strong>{{ item.summary }}</strong></h6>
                 <ul class="list-inline">
                     <li class="list-inline-item">
-                        <i class="fa fa-calendar" aria-hidden="true"></i> {{ item.start.strftime("%A") }}
+                        <i class="fa fa-calendar" aria-hidden="true"></i> {{ item.start | format_time("%A") }}
                     </li>
                     {% if item.type == "time" %}
                     <li class="list-inline-item">
-                        <i class="fa fa-clock" aria-hidden="true"></i> {{ item.start.strftime("%H:%M") }} - {{ item.end.strftime("%H:%M") }}
+                        <i class="fa fa-clock" aria-hidden="true"></i> {{ item.start | format_time("%H:%M") }} - {{ item.end | format_time("%H:%M") }}
                     </li>
                     {% endif %}
                     {% if item.location %}
