+++ flirror/views.py
@@ -1,11 +1,8 @@
 import abc
-from datetime import datetime
 
 import requests
 from flask import current_app, render_template, url_for
 from flask.views import MethodView
-
-from flirror.database import get_object_by_key
 
 
 class FlirrorMethodView(MethodView):
@@ -65,7 +62,6 @@ class WeatherView(FlirrorMethodView):
         res = requests.get(url_for("api-weather", _external=True))
         # TODO Error handling?
         weather = res.json()
-        print(weather)
 
         # Provide weather data in template context
         context = self.get_context(weather=weather)
@@ -78,31 +74,19 @@ class CalendarView(FlirrorMethodView):
     rule = "/calendar"
     template_name = "calendar.html"
 
-    FLIRROR_OBJECT_KEY = "module_calendar"
-
     def get(self):
         # Get view-specific settings from config
         # TODO Do we need to filter for these calendars?
         #  settings = current_app.config["MODULES"].get(self.endpoint)
         #  calendars = settings["calendars"]
 
-        data = self.get_events()
+        res = requests.get(url_for("api-calendar", _external=True))
+        # TODO Error handling?
+        data = res.json()
+
         # Provide events in template context
         context = self.get_context(**data)
         return render_template(self.template_name, **context)
-
-    def get_events(self):
-        db = current_app.extensions["database"]
-        data = get_object_by_key(db, self.FLIRROR_OBJECT_KEY)
-        # Change timestamps to datetime objects
-        # TODO This should be done before storing the data, but I'm not sure
-        # how to tell Pony how to serialize the datetime to JSON
-        for event in data["events"]:
-            event["start"] = datetime.utcfromtimestamp(event["start"])
-            event["end"] = datetime.utcfromtimestamp(event["end"])
-
-        data["_timestamp"] = datetime.utcfromtimestamp(data["_timestamp"])
-        return data
 
 
 class MapView(FlirrorMethodView):
